refactor: Assign values to types and files for reusability
