+++ frontend/src/lib/dnd/config.ts
@@ -1,0 +1,7 @@
+/**
+ * DnD entity type.
+ */
+export const DND_ENTITY_TYPE = {
+  ITEM: 'item',
+  COLUMN: 'column',
+} as const;
