+++ frontend/src/lib/dnd/entities/index.ts
@@ -1,3 +1,3 @@
+export * from './dnd-item';
 export * from './draggable-item';
 export * from './droppable-item';
-export * from './type';
