+++ frontend/src/lib/dnd/types.ts
@@ -1,0 +1,7 @@
+import type { DND_ENTITY_TYPE } from '@/lib/dnd/config';
+import type { ValueOf } from '@/types/utils';
+
+/**
+ * DnD entity type
+ */
+export type DndEntityType = ValueOf<typeof DND_ENTITY_TYPE>;
